requirements.txt
Flask-SQLAlchemy==2.0
Flask==0.10.1
Flask-Login==0.2.11
redis==2.10.5
Flask-Redis==0.1.0
MySQl-python
xlrd==1.0.0
xlwt==1.1.2
Flask-Script==2.0.5
Flask_RESTful==0.3.5
Pillow==3.4.2
baidu-aip==2.0.0.1
gevent==1.2.1
pymongo==3.4.0